Move tenant/delete.rs to tenant/timeline/delete.rs
